--- README.md
@@ -1,13 +1,0 @@
-## DEMO - https://crypto-trkr.netlify.app/
-
-
-## Getting Started
-#### `1. npm install`
-
-#### `2. npm run dev`
-
-<br>
-
-## TO DO:
-
-to be added later
