+++ src/components/Banner/MovingText.tsx
@@ -28,7 +28,6 @@ const MovingText = (props: Props) => {
   return (
     <svg
       className={styles['banner-svg']}
-      xmlns="http://www.w3.org/2000/svg"
       viewBox="0 15 200 500"
       preserveAspectRatio="xMinYMid meet"
       fill="currentColor"
