+++ src/layouts/Footer.tsx
@@ -1,10 +1,9 @@
 import styles from './Footer.module.css';
-import { useSelector } from 'react-redux';
+import { useAppSelector } from '../store/hooks';
 import { useEffect, useState } from 'react';
-import { RootState } from '../store';
 
 const Footer = () => {
-  const errorStatus = useSelector((state: RootState) => state.ui.hasError);
+  const errorStatus = useAppSelector((state) => state.ui.hasError);
   const [notificationStyle, setNotificationStyle] = useState({
     text: 'connected',
     style: `${styles.notification} ${styles.connected}`,
@@ -28,13 +27,6 @@ const Footer = () => {
     <footer className={styles.footer}>
       <div className={styles['footer-container']}>
         <div className={notificationStyle.style}>{notificationStyle.text}</div>
-        <div className={styles.credentials}>
-          <p>made by Popov Dmitry</p>
-
-          <a href="https://github.com/PopovDS617" target="blank">
-            GitHub
-          </a>
-        </div>
       </div>
     </footer>
   );
