+++ src/store/hooks.ts
@@ -1,0 +1,7 @@
+import { TypedUseSelectorHook, useSelector, useDispatch } from 'react-redux';
+
+import type { RootState, AppDispatch } from './store';
+
+export const useAppDispatch = () => useDispatch<AppDispatch>();
+
+export const useAppSelector: TypedUseSelectorHook<RootState> = useSelector;
